HW-4 Code Done
Code for HW-4 assignment is done. All test programs run to completion with
no errors. Changes have been made to CPU file as well - The PC is now
incremented at the end of the loop always, and for branch statements the
address that it branches to is now subtracted by 1 instruction size, to
account for the PC adding at the end.
Changes to SyscallOpen and SyscallClose structure also took place.
Method headers and Documenting has not been completed yet.
Still wondering about the rate at which the interleaving happens for the
test program. This will be looked into later.