Documentation Update and Small Code Update
Comments in about the program have been changed to reflect the changes in
code.
Schedule new process now checks if the selected new process has the same
ID as the current process, and if so, does not proform a context switch>
SyscallOpen now preforms a scheduleNewProcess call if the device that is
trying to open a device is blocked, in addition to a syscall Success.