HW-6 Mostly Done
All of the code for HW-6 seems to be done. There is one weird bug, in that
every once in a while, the program crashes and gives an array out of bounds
error. The CPU usage also spikes way up when this happens. We have no clue
why this is happening. Will look into later.
Also need to put in more comments.