HW-6 Final Commmit
Switched OS code from my original OS code to using Justice Nichols OS code
from previous assignments.
Code should be complete and everything running with the correct times and
what not as long as nothing in the code changes. There were a ton of problems
with the run times varrying vastly with changes to verbose mode and other
changes that would seemingly not effect the code.